+++ targets/C++.source/token.h
@@ -1,13 +1,15 @@
 /* Stack Token */
 typedef struct
 {
-	@@prefix_vtype			value;
-	@@prefix_ast*			node;
+	@@prefix_vtype		value;
+	@@prefix_ast*		node;
 
-	const @@prefix_syminfo*	symbol;
+	const
+	@@prefix_syminfo*	symbol;
 
-	int						state;
-	unsigned int			line;
-	unsigned int			column;
+	int					state;
+
+	size_t				line;
+	size_t				column;
 } @@prefix_tok;
